# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OmniMerchant API
  version: 0.2.0
  description: Spring AI cross-border ecommerce customer-service platform API. Static v2 contract for local demo and reviewer verification.
servers:
  - url: http://localhost:8090
security:
  - bearerAuth: []
paths:
  /api/health:
    get:
      summary: Health check
      security: []
      responses:
        '200': { description: OK }
  /api/admin/login:
    post:
      summary: Admin login
      security: []
      responses:
        '200': { description: JWT response }
  /api/widget/session:
    post:
      summary: Create public buyer widget session
      security: []
      responses:
        '200': { description: Widget session response with customerSessionToken and expiresAt }
  /api/widget/chat/stream:
    post:
      summary: Public buyer SSE chat
      security:
        - bearerAuth: []
      responses:
        '200': { description: text/event-stream }
        '401': { description: Missing, invalid, or expired customerSessionToken }
        '403': { description: Token tenant or conversation mismatch }
  /api/chat/stream:
    post:
      summary: Authenticated merchant test SSE chat
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200': { description: text/event-stream }
        '401': { description: Invalid JWT }
        '403': { description: Tenant membership mismatch }
  /api/customers:
    get:
      summary: List customers
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: keyword, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200': { description: Customer page }
  /api/orders:
    get:
      summary: List orders
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: keyword, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200': { description: Order page }
  /api/orders/by-number/{orderNumber}:
    get:
      summary: Get order by display number
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: orderNumber, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Order detail }
  /api/products:
    get:
      summary: List products
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: keyword, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
      responses:
        '200': { description: Product page }
  /api/products/reindex:
    post:
      summary: Mark products for vector reindex
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200': { description: Queued count }
  /api/escalations:
    get:
      summary: List escalation tickets
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: status, in: query, schema: { type: integer } }
      responses:
        '200': { description: Escalation page }
    post:
      summary: Create escalation ticket
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200': { description: Escalation detail }
  /api/escalations/{id}/assign:
    put:
      summary: Assign escalation
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: Escalation detail }
  /api/escalations/{id}/resolve:
    put:
      summary: Resolve escalation
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: Escalation detail }
  /api/tool-calls:
    get:
      summary: List tool audit records
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200': { description: Tool-call page }
  /api/dashboard/commerce:
    get:
      summary: Commerce dashboard metrics
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200': { description: Dashboard metrics }
  /api/integrations/shopify/connect:
    post:
      summary: Save encrypted Shopify credentials
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200': { description: Connection status }
  /api/integrations/shopify/install:
    get:
      summary: Start Shopify OAuth install
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: shop, in: query, required: true, schema: { type: string, example: example.myshopify.com } }
      responses:
        '200': { description: OAuth install URL and state }
        '400': { description: Invalid shop domain }
  /api/integrations/shopify/oauth/callback:
    get:
      summary: Shopify OAuth callback
      security: []
      parameters:
        - { name: shop, in: query, required: true, schema: { type: string } }
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
        - { name: hmac, in: query, required: true, schema: { type: string } }
      responses:
        '200': { description: OAuth token saved and sync jobs created }
        '401': { description: Invalid HMAC or state }
  /api/integrations/shopify/sync:
    post:
      summary: Sync Shopify customers, products, and orders
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200': { description: Sync summary }
  /api/integrations/shopify/jobs:
    get:
      summary: List Shopify cursor sync jobs
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200': { description: Sync job page }
  /api/integrations/shopify/jobs/{jobId}/retry:
    post:
      summary: Retry a failed Shopify sync job
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: jobId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: Updated job status }
        '404': { description: Job not found }
  /api/integrations/shopify/webhooks:
    get:
      summary: List Shopify webhook processing records
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200': { description: Webhook event page }
  /api/integrations/shopify/webhooks/{eventId}/replay:
    post:
      summary: Replay a stored Shopify webhook event
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: eventId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: Webhook replay result }
        '404': { description: Webhook event not found }
  /api/webhooks/shopify:
    post:
      summary: Shopify webhook receiver
      security: []
      responses:
        '200': { description: Webhook accepted or rejected with app error code }
  /api/evals:
    get:
      summary: List golden eval cases
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200': { description: Eval case page }
  /api/evals/run:
    post:
      summary: Run deterministic or opt-in live tenant evals
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200': { description: Eval run report }
  /api/evals/runs:
    get:
      summary: List persisted eval runs
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200': { description: Eval run page }
  /api/evals/runs/{runId}:
    get:
      summary: Get eval run detail and case results
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: runId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: Eval run detail }
  /api/observability/summary:
    get:
      summary: Tenant-scoped AI quality, cost, latency, and safety summary
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
      responses:
        '200': { description: Observability KPI summary }
  /api/observability/failures:
    get:
      summary: Failure attribution buckets
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: category, in: query, schema: { type: string } }
      responses:
        '200': { description: Failure bucket list }
  /api/observability/traces:
    get:
      summary: List agent traces
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: conversationUuid, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [RUNNING, SUCCESS, FAILED] } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200': { description: Trace page }
  /api/observability/traces/{traceId}:
    get:
      summary: Get trajectory replay details
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: traceId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Trace timeline }
        '404': { description: Trace not found }
  /api/rag/safety/docs:
    get:
      summary: List RAG ingestion safety reviews
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: status, in: query, schema: { type: string, enum: [APPROVED, QUARANTINED, REJECTED] } }
        - { name: riskLevel, in: query, schema: { type: string, enum: [LOW, MEDIUM, HIGH] } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200': { description: RAG safety review page }
  /api/rag/safety/docs/{docUuid}/approve:
    post:
      summary: Approve a reviewed document for indexing
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: docUuid, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Updated review }
  /api/rag/safety/docs/{docUuid}/reject:
    post:
      summary: Reject or quarantine a reviewed document
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: docUuid, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Updated review }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: integer
        format: int64